+++ CHANGELOG.rst
@@ -2,8 +2,8 @@
 Change Log
 ==========
 
-HEAD
-----
+1.9.3
+-----
 * Improved support for ``filter()``, ``.only()``, ``.order_by()`` etc. on indexed properties. It is now possible to
   specify labels and subfields, e.g. ``.filter(phone_numbers=PhoneNumber(label='CarPhone', phone_number='123'))``
   ``.filter(phone_numbers__CarPhone='123')``, ``.filter(physical_addresses__Home__street='Elm St. 123')``,